+++ libs/sdk-backend-spi/src/deprecated/userGroups.ts
@@ -1,0 +1,10 @@
+// (C) 2019-2022 GoodData Corporation
+/* eslint-disable @typescript-eslint/no-empty-interface, import/namespace */
+import * as m from "@gooddata/sdk-model";
+
+/**
+ * User Group
+ * @deprecated Use {@link @gooddata/sdk-model#IWorkspaceUserGroup}
+ * @alpha
+ */
+export interface IWorkspaceUserGroup extends m.IWorkspaceUserGroup {}
