+++ libs/sdk-backend-spi/src/workspace/userGroups/index.ts
@@ -1,29 +1,6 @@
 // (C) 2019-2022 GoodData Corporation
-import { ObjRef } from "@gooddata/sdk-model";
+import { IWorkspaceUserGroup } from "@gooddata/sdk-model";
 import { IPagedResource } from "../../common/paging";
-
-/**
- * User Group
- * @alpha
- */
-export interface IWorkspaceUserGroup {
-    /**
-     * Stored user group reference
-     */
-    ref: ObjRef;
-    /**
-     * Stored user group id
-     */
-    id?: string;
-    /**
-     * Group name
-     */
-    name?: string;
-    /**
-     * Group description
-     */
-    description?: string;
-}
 
 /**
  * Configuration options for querying user groups
